move ownership check into allocator base
